requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.2
boto3==1.28.30
botocore==1.31.30
bs4==0.0.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
crispy-bootstrap4==2022.1
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
Django==4.2.2
django-admin-persian-fonts==0.2
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-js-asset==2.1.0
django-rest-framework==0.1.0
django-restframework==0.0.1
django-storages==1.13.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
executing==1.2.0
filelock==3.10.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.21.3
ipython==8.11.0
jdatetime==4.1.1
jedi==0.18.2
jmespath==1.0.1
jupyter_client==8.0.3
jupyter_core==5.2.0
Khayyam==3.0.17
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
mysqlclient==2.2.0
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.0
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
static3==0.7.0
termcolor==2.3.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.21.0
wcwidth==0.2.6
